+++ css/common/common.css
@@ -4,6 +4,7 @@
     format("woff2");
 
 }
+
 * {
   margin: 0;
   padding: 0;
@@ -26,9 +27,11 @@ body,
   font-size: larger;
   font-family: 'Poppins', sans-serif;
 }
+
 .backspace a:link{
   text-decoration: none;
 }
+
 #header {
   font-family: "iceJaram-Rg";
   font-size: 9em;
