+++ css/feature/mainPage.css
@@ -6,7 +6,6 @@
   font-style: normal;
 }
 
-
 #mainpage_Container {
   width: 100%;
   height: 100vh;
@@ -119,7 +118,6 @@ a span:nth-child(2) {
       top: 100%;
   }
 }
-
 
 a span:nth-child(4) {
   bottom: -100%;
