+++ css/style.css
@@ -3,7 +3,6 @@
   justify-content: center;
 }
 
-
 .header-detail span {
   color: #FFF;
   font-family: Poppins;
