+++ html/feature/detailPage.html
@@ -152,9 +152,9 @@
         <!--footer 추가-->
         <div id="footer"></div>
     </div>
-    <script src="/js/common.js"></script>
-    <script src="/js/updatecomment.js"></script>
-    <script src="/js/detailPage.js"></script>
-    <script src="/js/modal.js"></script>
+    <script src="/js/common/common.js"></script>
+    <script src="/js/feature/updatecomment.js"></script>
+    <script src="/js/page/detailPage.js"></script>
+    <script src="/js/feature/modal.js"></script>
 </body>
 </html>
