+++ html/feature/mainPage.html
@@ -48,6 +48,6 @@
       </div>
     </div>
   </div>
-  <script type="module" src="../../js/moveMovieList.js"></script>
+  <script type="module" src="../../js/feature/moveMovieList.js"></script>
 </body>
 </html>
