+++ index.html
@@ -8,7 +8,6 @@
   <link rel="stylesheet" href="./css/common/common.css" />
   <link rel="stylesheet" href="./css/style.css" />
   <script src="https://kit.fontawesome.com/d84eab0825.js" crossorigin="anonymous"></script>
-  <!-- <script src="/js/common.js" type="text"></script> -->
 </head>
 
 <body>
@@ -36,15 +35,14 @@
     </div>
 
     <div id="card_list" class="movie-card-list">
-
     </div>
   
-    <!--footer 추가-->
     <div id="footer"></div>
   </div>
-  <script src="/js/common.js"></script>
-  <!-- <script src="./scripts/movie_list.js" type="module"></script> -->
-  <script src="./scripts/movie_list.js"></script>
-  <script src="./scripts/search.js"></script>
+  
+  <script src="./js/common/common.js"></script>
+  <script src="./js/page/movie_list.js"></script>
+  <script src="./js/feature/search.js"></script>
+
 </body>
 </html>
