+++ js/config/config.js
