+++ js/feature/moveMovieList.js
@@ -1,11 +1,8 @@
 const maintitle_Container = document.querySelector('#maintitle_Container');
+
 maintitle_Container.addEventListener('click', function (e) {
-    console.log('여기는 아이디 만드는곳.@@@@@@@@@@@@@@@@@@@@@@@@@@');
+    
     const buttonId = e.target.id;
-    console.log(buttonId);
-
-    console.log('여기는 카드 만드는곳');
-
     const options = {
         method: 'GET',
         headers: {
@@ -22,53 +19,59 @@ maintitle_Container.addEventListener('click', function (e) {
     const fetchData = async (urls) => {
         const response = await fetch(urls, options);
         const data = await response.json();
-        console.log(data)
         return data;
     };
     
-    // 각 URL에 대해 fetchData 함수를 호출하고 결과를 기다림
+    /**
+     * 각 URL에 대해 fetchData 함수를 호출하고 결과를 기다림
+     */
     const fetchPromises = urls.map(url => fetchData(url));
 
-// 모든 Promise가 처리되도록 기다림
-Promise.allSettled(fetchPromises)
-    .then(results => {
-        let newMovies = [];
-        let movies = [results[0]['value']['results'], results[1]['value']['results']];
-        for(let i = 0; i < movies.length; i++){
-            for(let j = 0; j<movies[0].length; j++){
-                newMovies.push(movies[i][j]);
+    /**
+     * 모든 Promise가 처리되도록 기다림
+     */
+    Promise.allSettled (fetchPromises)
+        .then(results => {
+            let newMovies = [];
+            let movies = [results[0]['value']['results'], results[1]['value']['results']];
+
+            for (let i = 0; i < movies.length; i++){
+                for (let j = 0; j<movies[0].length; j++){
+                    newMovies.push(movies[i][j]);
+                }
             }
-        }
 
-        let sortedMovies = sortMovie(newMovies,buttonId)
-        makeCard(sortedMovies)
-    })
-    .catch(error => {
-        console.error(error);
-    });
-
-    // return buttonId;
+            let sortedMovies = sortMovie(newMovies,buttonId)
+            makeCard(sortedMovies)
+        })
+        .catch(error => {
+            console.error(error);
+        });
 });
 
 
-function sortMovie(movies,buttonId){
-    if(buttonId == 'sortRatings'){
+function sortMovie (movies,buttonId){
+    if (buttonId == 'sortRatings') {
         return movies.sort((a, b) => b['vote_average'] - a['vote_average']);
-    }else if(buttonId == 'sortTitle'){
+    } else if (buttonId == 'sortTitle') {
         return movies.sort((a, b) => a['title'].localeCompare(b['title']));
-    }else if(buttonId == 'sortpopular'){
+    } else if (buttonId == 'sortpopular' ){
         return movies.sort((a, b) => b['popularity']-(a['popularity']));
-    }else if(buttonId == 'sortComming'){
-        return  movies.sort((a, b) => b['release_date'].localeCompare(a['release_date']));
+    } else if (buttonId == 'sortComming') {
+        return movies.sort((a, b) => b['release_date'].localeCompare(a['release_date']));
     }
 }
 
 
-function makeCard(movies) {
-    //sortfunc의 return값 사용할 위치
-    console.log('카드만드는곳.@@@@@@@@@@@@@@@@@@@@@@@@@@')
+/**
+ * sortfunc의 return값 사용할 위치
+ */
+function makeCard (movies) {
+
     const queryParams = new URLSearchParams();
+    
     for (let i = 0; i < movies.length; i++) {
+    
         let movie = movies[i]
         let movieInfo = {
             title: `${movie['title']}`,
@@ -78,10 +81,10 @@ function makeCard(movies) {
             popularity: `${movie['popularity']}`,
             release_date: `${movie['release_date']}`
         }
+    
         let movieInfoString = JSON.stringify(movieInfo)
+    
         window.localStorage.setItem(`${movie['id']}`, movieInfoString)
-
-        console.log('값 다 들어감.');
         queryParams.append('cardIds', movie['id']);
     }
     window.location.href = `../../index.html?${queryParams.toString()}`
