+++ js/page/detailPage.js
